+++ examples/main/shared/editor/EditorPlugins.tsx
@@ -870,7 +870,7 @@ const config: RichContentEditorProps['config'] = {
           ],
         },
       }),
-      // buttonsMapper: () => ({
+      // buttonsOverrides: () => ({
       //   desktop: {
       //     Bold: {
       //       icon: () => <div>bla</div>,