+++ packages/plugin-social-polls/web/src/tiptap/component.tsx
@@ -3,6 +3,7 @@ import { PollEditor } from '../components';
 import { POLL_TYPE } from '../types';
 import { PluginProps } from 'wix-rich-content-editor-common';
 import { convertBlockDataToRicos } from 'ricos-content/libs/migrateSchema';
+import { NodeSelection } from 'prosemirror-state';
 
 export const Poll: React.FC<PluginProps> = ({
   context,
@@ -10,6 +11,8 @@ export const Poll: React.FC<PluginProps> = ({
   node,
   updateAttributes,
   selected,
+  editor,
+  getPos,
 }) => {
   const { theme, t, config = {}, isMobile } = context;
   const settings = config[POLL_TYPE] || {};
@@ -23,18 +26,30 @@ export const Poll: React.FC<PluginProps> = ({
   };
   const setInPluginEditingMode = () => {};
 
+  const setSelection = () => {
+    if (!selected) {
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const selection = NodeSelection.create(editor.view.state.doc, (getPos as any)());
+      const transaction = editor.view.state.tr.setSelection(selection);
+      editor.view.dispatch(transaction);
+    }
+  };
+
   return (
-    <PollEditor
-      componentData={componentData}
-      settings={settings}
-      theme={theme}
-      t={t}
-      isMobile={isMobile}
-      block={block}
-      selected={selected}
-      helpers={helpers}
-      store={store}
-      setInPluginEditingMode={setInPluginEditingMode}
-    />
+    // eslint-disable-next-line jsx-a11y/no-static-element-interactions
+    <div onMouseDown={setSelection}>
+      <PollEditor
+        componentData={componentData}
+        settings={settings}
+        theme={theme}
+        t={t}
+        isMobile={isMobile}
+        block={block}
+        selected={selected}
+        helpers={helpers}
+        store={store}
+        setInPluginEditingMode={setInPluginEditingMode}
+      />
+    </div>
   );
 };