+++ packages/plugin-spoiler/web/src/tiptap/tiptap.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
-import { CreateRicosExtensions } from 'ricos-tiptap-types';
+import { CreateRicosExtensions, DOMOutputSpec } from 'ricos-tiptap-types';
 import { BlockSpoilerComponent } from '..';
+import colorDataDefaults from 'ricos-schema/dist/statics/color.defaults.json';
+
+const SPOILER_STYLE = 'blur(0.25em)';
 
 const SpoilerHoc = Component => {
   // should use the new api containerData
@@ -30,11 +33,22 @@ const SpoilerHoc = Component => {
   return Spoiler;
 };
 
+declare module '@tiptap/core' {
+  interface Commands<ReturnType> {
+    spoiler: {
+      /**
+       * Set a spoiler mark
+       */
+      toggleSpoiler: () => ReturnType;
+    };
+  }
+}
+
 export const createTiptapExtensions: CreateRicosExtensions = defaultOptions => [
   {
     type: 'extension' as const,
     createExtensionConfig: () => ({
-      name: 'spoiler',
+      name: 'node-spoiler',
       priority: 10,
       addOptions: () => defaultOptions,
       addNodeHoc: () => ({
@@ -42,6 +56,44 @@ export const createTiptapExtensions: CreateRicosExtensions = defaultOptions => [
         priority: 10,
         nodeHoc: SpoilerHoc,
       }),
+    }),
+  },
+  {
+    type: 'mark' as const,
+    createExtensionConfig: () => ({
+      name: 'spoiler',
+
+      addOptions: () => ({
+        HTMLAttributes: {},
+      }),
+
+      addAttributes() {
+        return colorDataDefaults;
+      },
+
+      parseHTML() {
+        return [
+          {
+            tag: 'span',
+            getAttrs: element => {
+              const { filter } = (element as HTMLElement).style || {};
+              return filter === SPOILER_STYLE ? {} : false;
+            },
+          },
+        ];
+      },
+
+      renderHTML() {
+        return ['span', { style: `filter: ${SPOILER_STYLE}` }, 0] as DOMOutputSpec;
+      },
+
+      addCommands() {
+        return {
+          toggleSpoiler: () => ({ commands }) => {
+            return commands.toggleMark('spoiler');
+          },
+        };
+      },
     }),
   },
 ];