+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/core-configs.ts
@@ -1,10 +1,16 @@
 import { RicosExtension } from 'ricos-tiptap-types';
+import { createCommandsConfig } from '../../core-commands';
 import {
   createDraftConfig,
   createFocusConfig,
   createHistoryConfig,
   createStylesConfig,
   createUniqueId,
+  createBulletedList,
+  createBold,
+  createParagraph,
+  createItalic,
+  createUnderline,
 } from 'wix-tiptap-extensions';
 
 export const coreConfigs: RicosExtension[] = [
@@ -13,4 +19,10 @@ export const coreConfigs: RicosExtension[] = [
   createStylesConfig(),
   createFocusConfig(),
   createUniqueId(),
+  createCommandsConfig(),
+  createBulletedList(),
+  createBold(),
+  createParagraph(),
+  createItalic(),
+  createUnderline(),
 ];